+++ src/meta/links/types.ts
@@ -1,0 +1,7 @@
+export interface IconMeta {
+  color?: string
+  fill?: boolean
+  extraBc?: string
+  extraBcRad?: string
+  svg: any
+}
